Update README.md
A static front-end website for enjoying leisure time with practicing Drums of a Band. It is built using fundamentals of pure HTML, CSS, JavaScript and jQuery, and some JS
frameworks. The website also has Event handing as well as audio features. The peculiarity of this project is I have implemented both features on KeyPress as well as OnMouseClick events . This project is an audio-based project and helps you to soothe your mind and will be more interesting if you are interested in listening to Band Music.